+++ futures_portfolio/tests/test_import.py
@@ -1,2 +1,0 @@
-import rank_tickers
-print('import ok')